% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{coxsnell_plot}
\alias{coxsnell_plot}
\title{Cox-Snell residual diagnostic plot}
\usage{
coxsnell_plot(object, ...)
}
\arguments{
\item{object}{a fitted \code{"rpsurv"} object.}

\item{...}{further arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Plots the Nelson-Aalen cumulative hazard of the Cox-Snell residuals
against the residuals themselves. A well-fitting model gives points
scattered around the y = x line.
}
