% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsurv-package.R
\docType{package}
\name{rpsurv-package}
\alias{rpsurv-package}
\title{rpsurv: Fast Royston-Parmar Flexible Parametric Survival Models}
\description{
Fits Royston-Parmar flexible parametric survival models with the
likelihood and gradient evaluated in parallel C++ ('RcppParallel').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ielbadisy/rpsurv}
  \item Report bugs at \url{https://github.com/ielbadisy/rpsurv/issues}
}

}
\author{
\strong{Maintainer}: Imad El Badisy \email{elbadisyimad@gmail.com}

}
\keyword{internal}
